pyproject.toml
[project]
name = "teresa-pn-ou5"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "numpy",
    "scipy",
    "pandas",
    "matplotlib>=3.3",
    "seaborn",
    "astropy",
    "cmasher",
    "pyyaml",
    "jupyter_client",
    "jupyter",
    "jupyterlab",
    "jupytext",
    "mes-longslit",
    "statsmodels>=0.14.1",
    "ipython<9", #2025-07-01: bug in 9.0.0
    "pyneb>=1.1.28",
    "cloudytab",
    "typer>=0.17.4",
    "pyaml>=25.7.0",
]

[project.optional-dependencies]
dev = [
    "flit",
    "black",
    "mypy"
]

[tool.uv.sources]
mes-longslit = { workspace = true }
cloudytab = { path = "../cloudytab", editable = true }

[tool.uv.workspace]
members = ["mes-longslit"]

[project.urls]
Home = "https://github.com/will-henney/teresa-pn-ou5"
